Refactor: 패키지 변경 (product_recommend -> admin)
