+++ src/main/java/com/ecolink/core/admin/constant/RecommendType.java
@@ -1,0 +1,5 @@
+package com.ecolink.core.admin.constant;
+
+public enum RecommendType {
+	TODAY, MY_PRODUCT
+}
