--- src/main/java/com/ecolink/core/product_recommend/constant/RecommendType.java
@@ -1,5 +1,0 @@
-package com.ecolink.core.product_recommend.constant;
-
-public enum RecommendType {
-	TODAY, MY_PRODUCT
-}
